server/tronbyt-webp-png/app/requirements.txt
flask
requests
pillow
imageio
numpy

